backend/onvif/requirements/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
colorama==0.4.6
beautifulsoup4==4.12.3
hydra-core==1.3.2
lxml==5.2.2
pika==1.3.2
opencv-python
imageio